The version has been updated to 0.0.0.5. 1. The clip queue is no longer updated if the application is out of focus. 2. Code reorganization, folder reorganization. 3. Now the Action Sound Pocket Manager receives pockets from the object, they no longer need to be assigned separately. 4. Interaction is simplified: to start using the sound system, it is enough to add the “Sound” prefab to the stage, set the desired clips and subscribe to the necessary events, if required.
