--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Docs.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7a1cb9d3d3dda374f88ec1f8c4310b9c
-folderAsset: yes
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
