--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Docs/API Reference.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: b0cf4d8ed8e512a4280381c121a7c11e
-folderAsset: yes
-timeCreated: 1543580589
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
