--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Docs/Manual.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 9d7d00a8ef9f90b488fd53534129a7ad
-folderAsset: yes
-timeCreated: 1543580589
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
