--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Docs/Manual/EazySoundManager-Manual.pdf.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c942081f573a221439211bdb5b27b1d1
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
