--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Docs/Version/Version.txt
@@ -1,2 +1,0 @@
-Current sound system version: 0.0.0.4
-Current plugin version: 1.3.1
