--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Scripts/ActionSoundPocketManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 990360a4842377244bacc711b3768c98
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
