--- Assets/Lineri.SoundSystem/Eazy Sound Manager/Scripts/SoundPocket.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b84d0858374119f4f856f194340fde75
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
