+++ Assets/Lineri/SoundSystem/Eazy Sound Manager/Docs/API Reference/EazySoundManager-APIReference.chm.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0050e7f214aa2c343b1ecfaf9a5877a5
+timeCreated: 1543580589
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
