+++ Assets/Lineri/SoundSystem/Eazy Sound Manager/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5aa6c2b5e7d16af49affa6f150da5edb
+folderAsset: yes
+timeCreated: 1471688415
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
