+++ Assets/Lineri/SoundSystem/Eazy Sound Manager/Scripts/Audio.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: fb53b2b01d9c011488ad229895b665b5
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
