+++ Assets/Lineri/SoundSystem/Eazy Sound Manager/Scripts/EazySoundManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0500072cba35d4749ad862e7730ad14b
+timeCreated: 1471688400
+licenseType: Store
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
