+++ Assets/Lineri/SoundSystem/SoundSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5d23870228a3434db9de426e12ed0f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
