+++ Assets/Lineri/SoundSystem/SoundSystem/Docs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c36d65f41ecff5642b198a55a8d78d3b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
