+++ Assets/Lineri/SoundSystem/SoundSystem/Docs/Version/Version.txt
@@ -1,0 +1,2 @@
+Current sound system version: 0.0.0.5
+Current plugin version: 1.3.1
