+++ Assets/Lineri/SoundSystem/SoundSystem/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e60a67f8aa8bb7408bc5da7881ffdcf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
