+++ Assets/Lineri/SoundSystem/SoundSystem/Prefabs/Sound.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dd226579b7d3d0748b9c2161182c02cd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
