+++ Assets/Lineri/SoundSystem/SoundSystem/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b14d854c2c3985a4385849087128dcaf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
