+++ Assets/Lineri/SoundSystem/SoundSystem/Scripts/ActionSoundPocketManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 990360a4842377244bacc711b3768c98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
